% Generated by roxygen2 (4.0.1): do not edit by hand
\name{convert.list.to.type}
\alias{convert.list.to.type}
\title{Converts a list to a type, by copying just the fields that the type knows of}
\usage{
convert.list.to.type(li, typeName, types = get.types())
}
\description{
Converts a list to a type, by copying just the fields that the type knows of
}

