% Generated by roxygen2 (4.0.1): do not edit by hand
\name{custom.whisker.render}
\alias{custom.whisker.render}
\title{Adapts whisker render for different whisker formats}
\usage{
custom.whisker.render(template, data, ..., whiskers = c("<<", ">>"))
}
\description{
Adapts whisker render for different whisker formats
}

