% Generated by roxygen2 (4.0.1): do not edit by hand
\name{deduce.typeName}
\alias{deduce.typeName}
\title{Returns the type of a field. Uses the definition in parent.type to derive the field type}
\usage{
deduce.typeName(tree.obj, name, parent, parent.type = get.type(parent))
}
\description{
Returns the type of a field. Uses the definition in parent.type to derive the field type
}

