% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.parent.field.row}
\alias{get.parent.field.row}
\title{Gets the closest parent field that is a subtype of supertype}
\usage{
get.parent.field.row(row, st, supertype)
}
\description{
Gets the closest parent field that is a subtype of supertype
}

