% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.parents.rows}
\alias{get.parents.rows}
\title{Gets all rows of objects that are structure parents of the current objects}
\usage{
get.parents.rows(row, st)
}
\description{
Gets all rows of objects that are structure parents of the current objects
}

