% Generated by roxygen2 (4.0.1): do not edit by hand
\name{is.condition}
\alias{is.condition}
\title{Parsing and assessing experiment and game structures}
\usage{
is.condition(obj, typeName = get.typeName(obj), types = get.types())
}
\description{
Parsing and assessing experiment and game structures
}

