% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.struct}
\alias{load.struct}
\title{Load a game, experiment or dataLink structure}
\usage{
load.struct(name, file = paste0(default.struct.path(), "/", name, ".yaml"),
  typeName = "game", just.obj = FALSE, types = get.types(), ...)
}
\description{
Load a game, experiment or dataLink structure
}

