% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.types}
\alias{load.types}
\title{Reads information about action types}
\usage{
load.types(type.files = paste0(default.types.path(), "/",
  list.files(default.types.path())), store.as.default = TRUE,
  high.prio.types = list(), low.prio.types = list())
}
\description{
Reads information about action types
}

