% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.object.from.type}
\alias{make.object.from.type}
\title{Generates an empty list that has all fields of a type}
\usage{
make.object.from.type(typeName, types = get.types())
}
\description{
Generates an empty list that has all fields of a type
}

