% Generated by roxygen2 (4.0.1): do not edit by hand
\name{obj.to.struct}
\alias{obj.to.struct}
\title{Retrieves for an object a structure consisting of an obj.li containing all descendants and a data frame df that contains some summary information}
\usage{
obj.to.struct(obj, name)
}
\description{
Retrieves for an object a structure consisting of an obj.li containing all descendants and a data frame df that contains some summary information
}

