% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.yaml}
\alias{read.yaml}
\title{Reads a yaml file and returns as a list}
\usage{
read.yaml(file = NULL, verbose = FALSE, quote.char = "__QUOTE__",
  text = NULL, catch.error = TRUE, check.by.row = FALSE)
}
\description{
Reads a yaml file and returns as a list
}

