% Generated by roxygen2 (4.0.1): do not edit by hand
\name{set.object.defaults}
\alias{set.object.defaults}
\title{Assign default values for non-specified fields of a game object (action etc...)}
\usage{
set.object.defaults(obj, type, fields = type$fields)
}
\description{
Assign default values for non-specified fields of a game object (action etc...)
}

