% Generated by roxygen2 (4.0.1): do not edit by hand
\name{str.starts.with}
\alias{str.starts.with}
\title{An operator that is true if the string str starts with the substring key}
\usage{
\method{str}{starts.with}(str, key)
}
\description{
An operator that is true if the string str starts with the substring key
}

