% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tree.obj.to.struct.obj}
\alias{tree.obj.to.struct.obj}
\title{Recursively parse a structure object of an experiment or game}
\usage{
tree.obj.to.struct.obj(tree.obj, name, typeName = NULL, parent = NULL,
  types = get.types())
}
\description{
Recursively parse a structure object of an experiment or game
}

